LayoutTests/platform/qt/fast/encoding/frame-default-enc-expected.txt



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
ISO-8859-1
The above line should be your browser's default encoding, and not x-mac-cyrillic.
